+++ src/squoosh-lib.d.ts
