--- src/utils/imageminMinify.js
@@ -1,167 +1,0 @@
-// import path from "path";
-import { klona } from "klona/full";
-// import fileTypeFromBuffer from "./fileTypeFromBuffer";
-
-/** @typedef {import("../index").ImageminMinimizerOptions} ImageminMinimizerOptions */
-/** @typedef {import("../index").WorkerResult} WorkerResult */
-/** @typedef {import("imagemin").Options} ImageminOptions */
-
-/**
- * @typedef {Object} MetaData
- * @property {Array<Error>} warnings
- * @property {Array<Error>} errors
- */
-
-class InvalidConfigError extends Error {
-  /**
-   * @param {string | undefined} message
-   */
-  constructor(message) {
-    super(message);
-
-    this.name = "InvalidConfigError";
-  }
-}
-
-/**
- * @param {ImageminMinimizerOptions} minimizerOptions
- */
-function imageminNormalizeConfig(minimizerOptions) {
-  if (
-    !minimizerOptions ||
-    !minimizerOptions.plugins ||
-    (minimizerOptions.plugins && minimizerOptions.plugins.length === 0)
-  ) {
-    throw new Error(
-      "No plugins found for `imagemin`. Please read documentation."
-    );
-  }
-
-  const imageminConfig = klona(minimizerOptions);
-
-  if (!imageminConfig.pluginsMeta) {
-    imageminConfig.pluginsMeta = [];
-  }
-
-  imageminConfig.plugins = imageminConfig.plugins.map((plugin) => {
-    const isPluginArray = Array.isArray(plugin);
-
-    if (typeof plugin === "string" || isPluginArray) {
-      const pluginName = isPluginArray
-        ? /** @type {[string, object]} */ (plugin)[0]
-        : /** @type {string} */ (plugin);
-      const pluginOptions = isPluginArray
-        ? /** @type {[string, object]} */ (plugin)[1]
-        : undefined;
-
-      let requiredPlugin = null;
-      let requiredPluginName = `imagemin-${pluginName}`;
-
-      try {
-        // eslint-disable-next-line import/no-dynamic-require
-        requiredPlugin = require(requiredPluginName)(pluginOptions);
-      } catch {
-        requiredPluginName = pluginName;
-
-        try {
-          // eslint-disable-next-line import/no-dynamic-require
-          requiredPlugin = require(requiredPluginName)(pluginOptions);
-        } catch {
-          const pluginNameForError = pluginName.startsWith("imagemin")
-            ? pluginName
-            : `imagemin-${pluginName}`;
-
-          throw new Error(
-            `Unknown plugin: ${pluginNameForError}\n\nDid you forget to install the plugin?\nYou can install it with:\n\n$ npm install ${pluginNameForError} --save-dev\n$ yarn add ${pluginNameForError} --dev`
-          );
-        }
-        // Nothing
-      }
-
-      let version = "unknown";
-
-      try {
-        // eslint-disable-next-line import/no-dynamic-require
-        ({ version } = require(`${requiredPluginName}/package.json`));
-      } catch {
-        // Nothing
-      }
-
-      /** @type {Array<Object>} imageminConfig.pluginsMeta */
-      (imageminConfig.pluginsMeta).push([
-        {
-          name: requiredPluginName,
-          options: pluginOptions || {},
-          version,
-        },
-      ]);
-
-      return requiredPlugin;
-    }
-
-    throw new InvalidConfigError(
-      `Invalid plugin configuration '${JSON.stringify(
-        plugin
-      )}', plugin configuration should be 'string' or '[string, object]'"`
-    );
-  });
-
-  return imageminConfig;
-}
-
-/**
- * @param {WorkerResult} original
- * @param {ImageminMinimizerOptions} options
- * @returns {Promise<WorkerResult>}
- */
-async function imageminMinify(original, options) {
-  // Implement autosearch config on root directory of project in future
-  const minimizerOptionsNormalized = /** @type {ImageminOptions} */ (
-    imageminNormalizeConfig(options)
-  );
-
-  const imagemin = require("imagemin");
-
-  let result;
-
-  try {
-    result = await imagemin.buffer(original.data, minimizerOptionsNormalized);
-  } catch (error) {
-    original.errors.push(
-      error instanceof Error ? error : new Error(/** @type {string} */ (error))
-    );
-
-    return original;
-  }
-
-  // const extInput = path.extname(original.filename).slice(1).toLowerCase();
-  // const { ext: extOutput } = fileTypeFromBuffer(result) || {};
-  //
-  // if (extOutput && extInput !== extOutput) {
-  //   original.warnings.push(
-  //     new Error(
-  //       `"imageminMinify" function do not support generate to "${extOutput}" from "${original.filename}". Please use "imageminGenerate" function.`
-  //     )
-  //   );
-  //
-  //   return original;
-  // }
-
-  return {
-    filename: original.filename,
-    data: result,
-    warnings: [...original.warnings],
-    errors: [...original.errors],
-    info: {
-      ...original.info,
-      minimized: true,
-      minimizedBy:
-        original.info && original.info.minimizedBy
-          ? ["imagemin", ...original.info.minimizedBy]
-          : ["imagemin"],
-    },
-  };
-}
-
-export default imageminMinify;
-export { imageminNormalizeConfig };
